首页页面查询接口;web访问页面查询和删除接口;评论页面查询和删除接口;博客管理页面查询接口

+++ mango_base/service_security/src/main/java/com/myblog/service/security/service/impl/RoleServiceImpl.java
@@ -135,7 +135,10 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
         List<String> delSuccessedRoleIds = new ArrayList<>();
         for (String id : ids) {
             Role deleteRole = baseMapper.selectById(id);
-            validRoleLevel(deleteRole.getLevel(), deleteRole.getRoleName());
+            Response response = validRoleLevel(deleteRole.getLevel(), deleteRole.getRoleName());
+            if (!response.getSuccess()) {
+                return response;
+            }
             List<String> adminIdsByRoleId = baseMapper.getAdminIdsByRoleId(id);
             if (!CollectionUtils.isEmpty(adminIdsByRoleId)) {
                 Role role = baseMapper.selectById(id);
@@ -169,7 +172,10 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
     @Transactional(rollbackFor = Exception.class)
     public Response editRole(RoleDto roleDto) throws Exception{
         Role dbRole = baseMapper.selectById(roleDto.getId());
-        validRoleLevel(dbRole.getLevel(), dbRole.getRoleName());
+        Response response = validRoleLevel(dbRole.getLevel(), dbRole.getRoleName());
+        if (!response.getSuccess()) {
+            return response;
+        }
         Role role = this.toDb(roleDto, Role.class);
         if (baseMapper.updateById(role) < 1) {
             LOGGER.error("editRole failed by unknown error, role:{}", roleDto);
@@ -187,7 +193,10 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
     @Transactional(rollbackFor = Exception.class)
     public Response updateMenu(RoleDto roleDto) throws Exception{
         Role dbRole = baseMapper.selectById(roleDto.getId());
-        validRoleLevel(dbRole.getLevel(), dbRole.getRoleName());
+        Response response = validRoleLevel(dbRole.getLevel(), dbRole.getRoleName());
+        if (!response.getSuccess()) {
+            return response;
+        }
         // 先删掉该角色绑定的所有菜单的信息
         baseMapper.deleteRoleMenuByRoleId(roleDto.getId());
 
@@ -230,31 +239,32 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
      * @param level
      * @return
      */
-    public void validRoleLevel(Integer level, String roleName) {
+    public Response validRoleLevel(Integer level, String roleName) {
         if (level == null) {
-            throw new RuntimeException("找不到被操作角色的级别");
+            return Response.error().message("找不到被操作角色的级别");
         }
         List<Role> roles = baseMapper.getRolesByUserId(SecurityUtils.getCurrentUserId());
         if (CollectionUtils.isEmpty(roles)) {
-            throw new RuntimeException("当前用户未绑定角色");
+            return Response.error().message("当前用户未绑定角色");
         }
         Set<String> roleNames = roles.stream().map(Role::getRoleName).collect(Collectors.toSet());
         // 如果所绑定的角色与被操作的角色名称相同，那么可以对其进行操作
         if (roleNames.contains(roleName)) {
-            return;
+            return Response.ok();
         }
         int currentLevel = roles.stream().mapToInt(Role::getLevel).min().getAsInt();
         if (currentLevel > level) {
-            throw new RuntimeException("权限不足，你的角色级别：" + currentLevel + "，低于操作的角色级别：" + level);
+            return Response.error().message("权限不足，你的角色级别：" + currentLevel + "，低于操作的角色级别：" + level);
         }
+        return Response.ok();
     }
 
-    public void validRoleLevelByUserId(String userId) {
+    public Response validRoleLevelByUserId(String userId) {
         List<Role> roles = baseMapper.getRolesByUserId(userId);
         Role role = roles.stream().min(Comparator.comparingInt(Role::getLevel)).orElse(null);
         if (Objects.isNull(role)) {
             throw new RuntimeException("当前用户未绑定角色");
         }
-        validRoleLevel(role.getLevel(), role.getRoleName());
+        return validRoleLevel(role.getLevel(), role.getRoleName());
     }
 }
